Adjustments made as requested
- Left changes to lines 52, 56, & 7
- Reverted changes to lines 69, 106, & 129
- Never changed line 255 - not sure why that one was flagged

Thanks for reviewing these! It's great to know the reasoning behind rejections. Is there a guide somewhere that lists out the Microsoft documentation standards such as the requirement for using present tense for localization? If so, that would be great to have a link to that prepopulated in the pull request to avoid wasting your time. Thanks again!