Create configure-exo-rules_protected.md
This document applies to Azure Information Protection labels that enforce access restrictions. The process of identifying labels in metadata of documents and headers of email is different when the labels enforce restriction as these are wrapped with encryption. I was not able to reproduce the behavior described in configure-exo-rules, but did not want to exclude the article as of yet.