Update requirements-applications.md
Hi,
Here is the link supporting the file formats in https://secude.com/halocad. You can see all the supported file formats in the bottom of the HALOCAD page in FAQ section. Here is also a recent post on HALOCAD published in Microsoft security blog which highlights the various supported CAD platforms
https://www.microsoft.com/security/blog/2021/07/22/how-to-protect-your-cad-data-files-with-mip-and-halocad/