Update quick-app-initialization-cpp.md
I was able to get the the [sensitivity labeling quickstart](https://docs.microsoft.com/en-us/information-protection/develop/quick-file-list-labels-cpp) up and running, but this required making a few changes to the boilerplate code in this section. Though I could be wrong, I believe the code actually won't compile as it's presented.

The changes I had to make:

* Changed `mAppInfo` reference to `appInfo`
* Removed `authDelegateImpl` from the `profileSettings` constructor
* Added `authDelegateImpl` to the `engineSettings` constructor
* Changed `setEngineId` to `SetEngineId`
* Changed `mipContext.Shutdown();` to `mMipContext->ShutDown();`
* Changed `mipContext = nullptr;` to `mMipContext = nullptr;`

Of course, feel free to let me know if the changes I'm proposing are symptomatic of an mistake on my end.